update refs to notificator redux submodule
