+++ src/modules/app/components/navbar/components/bell/component/bell.tsx
@@ -11,8 +11,7 @@ import "./bell.dark.scss";
 import { NotificationHistoryItem } from "@app/models";
 
 import { CompositeAppState } from "@app/redux";
-
-import { setNotificationsBadgeToZero } from "@app/redux/notificator/actions/setNotificationsBadgeToZero";
+import { setNotificationsBadgeToZero } from "@app/redux/notificator";
 
 import BellList from "../components/bellList/bellList";
