+++ src/redux/comments/thunks/increment.ts
@@ -4,8 +4,7 @@ import { post } from "@app/utils";
 
 import { ActionWithPayload, CompositeAppState } from "@app/redux";
 import { setError, getSetAppIsLoadingAction } from "@app/redux/app";
-
-import { getSuccessNotificationAction } from "@app/redux/notificator/utils";
+import { getSuccessNotificationAction } from "@app/redux/notificator";
 
 import { getIncrementAction } from "../actions/increment";
