+++ src/redux/comments/thunks/updateComment.ts
@@ -6,10 +6,9 @@ import { get, post } from "@app/utils";
 
 import { ActionWithPayload, CompositeAppState } from "@app/redux";
 import { setError, getSetAppIsLoadingAction } from "@app/redux/app";
+import { getSuccessNotificationAction } from "@app/redux/notificator";
 
 import { getOpenModalAction } from "@app/redux/modal/actionCreators/open";
-
-import { getSuccessNotificationAction } from "@app/redux/notificator/utils";
 
 import { getCommentModalFormCallbackConfig, getCommentModalFormConfig } from "../utils";
 import { getUpdateCommentAction } from "../actions/updateComment";
