+++ src/redux/notificator/actionCreators/index.ts
@@ -1,0 +1,4 @@
+export * from "./addNotification";
+export * from "./hideAllNotifications";
+export * from "./hideNotification";
+export * from "./setNotificationsBadgeToZero";
