+++ src/redux/notificator/actions.ts
@@ -1,11 +1,11 @@
 /** Add notification to store */
-export const AddNotification = "notification/add";
+export const ADD_NOTIFICATION = "notification/add";
 
 /** Hide specified notifications */
-export const HideNotification = "notification/hide";
+export const HIDE_NOTIFICATION = "notification/hide";
 
 /** Clear all active notifications */
-export const HideAllNotifications = "notification/hideAll";
+export const HIDE_ALL_NOTIFICATIONS = "notification/hideAll";
 
 /** Clear notification badge by setting to zero */
-export const SetNotificationsBadgeToZero = "notification/setBadgeToZero";
+export const SET_NOTIFICATIONS_BADGE_TO_ZERO = "notification/setBadgeToZero";
