--- src/redux/notificator/actions/hideAllNotifications.ts
@@ -1,11 +1,0 @@
-import { Action } from "@app/redux/types";
-
-import { HideAllNotifications } from "../actions";
-
-/**
- * Get redux action "Hide all notifications"
- * @returns State updating action
- */
-export const getHideAllNotificationsAction = (): Action => ({
-    type: HideAllNotifications
-});
