+++ src/redux/notificator/index.ts
@@ -1,0 +1,4 @@
+export * from "./actionCreators";
+export * from "./actions";
+export * from "./types";
+export * from "./utils";
