+++ src/redux/notificator/reducer.ts
@@ -1,13 +1,11 @@
 import { getPropertyValueWithCheck } from "@bodynarf/utils";
 
 import { NotificationItem } from "@app/models";
-import { NotificatorState } from "./types";
 
 import { removeByKey } from "@app/utils";
 
-
-import { AddNotification, HideAllNotifications, HideNotification, SetNotificationsBadgeToZero } from "./actions";
 import { ActionWithPayload } from "@app/redux";
+import { NotificatorState, ADD_NOTIFICATION, HIDE_ALL_NOTIFICATIONS, HIDE_NOTIFICATION, SET_NOTIFICATIONS_BADGE_TO_ZERO } from "@app/redux/notificator";
 
 /** Default state of notification module */
 const defaultState: NotificatorState = {
@@ -19,7 +17,7 @@ const defaultState: NotificatorState = {
 /** Notification redux reducer function */
 export default function (state: NotificatorState = defaultState, action: ActionWithPayload): NotificatorState {
     switch (action.type) {
-        case AddNotification: {
+        case ADD_NOTIFICATION: {
             const addingNotifications: Array<NotificationItem> = getPropertyValueWithCheck(action.payload, "notifications", false);
 
             if (addingNotifications.length === 0) {
@@ -47,7 +45,7 @@ export default function (state: NotificatorState = defaultState, action: ActionW
                 historyBadgeCount: historyBadgeCount,
             };
         }
-        case HideNotification: {
+        case HIDE_NOTIFICATION: {
             const addingNotifications: Array<string> = getPropertyValueWithCheck(action.payload, "notificationIds", false);
 
 
@@ -61,13 +59,13 @@ export default function (state: NotificatorState = defaultState, action: ActionW
                 notifications: removeByKey(state.notifications, x => x.id, addingNotifications),
             };
         }
-        case HideAllNotifications: {
+        case HIDE_ALL_NOTIFICATIONS: {
             return {
                 ...state,
                 notifications: [],
             };
         }
-        case SetNotificationsBadgeToZero: {
+        case SET_NOTIFICATIONS_BADGE_TO_ZERO: {
             return {
                 ...state,
                 historyBadgeCount: 0,
