+++ src/redux/notificator/utils.ts
@@ -2,8 +2,8 @@ import { generateGuid } from "@bodynarf/utils";
 
 import { NotificationType } from "@app/models";
 
-import { getAddNotificationAction } from "./actions/addNotification";
 import { ActionWithPayload } from "@app/redux";
+import { getAddNotificationAction } from "@app/redux/notificator";
 
 /**
  * Get notifications module action which adding success notification
@@ -14,7 +14,16 @@ import { ActionWithPayload } from "@app/redux";
 export const getSuccessNotificationAction = (
     message: string,
     shouldDisplay: boolean
-): ActionWithPayload => getAddNotificationAction({ type: NotificationType.success, message, id: generateGuid(), createdAt: new Date() }, shouldDisplay);
+): ActionWithPayload =>
+    getAddNotificationAction(
+        {
+            type: NotificationType.success,
+            message,
+            id: generateGuid(),
+            createdAt: new Date()
+        },
+        shouldDisplay
+    );
 
 /**
  * Get notifications module action which adding error notification
@@ -25,4 +34,13 @@ export const getSuccessNotificationAction = (
 export const getErrorNotificationAction = (
     message: string,
     shouldDisplay: boolean
-): ActionWithPayload => getAddNotificationAction({ type: NotificationType.error, message, id: generateGuid(), createdAt: new Date() }, shouldDisplay);
+): ActionWithPayload =>
+    getAddNotificationAction(
+        {
+            type: NotificationType.error,
+            message,
+            id: generateGuid(),
+            createdAt: new Date()
+        },
+        shouldDisplay
+    );
